test : 쉐어 상세정보 조회 API 실패케이스 통합 테스트 구현
