+++ src/test/java/louie/hanse/shareplate/integration/share/ShareIntegrationTest.java
@@ -70,49 +70,6 @@ class ShareIntegrationTest extends InitIntegrationTest {
     }
 
     @Test
-    void 특정_쉐어의_상세정보를_조회한다() {
-        given(documentationSpec)
-            .filter(document("share-detail-get"))
-            .accept(APPLICATION_JSON_VALUE)
-            .pathParam("id", 1)
-
-            .when()
-            .get("/shares/{id}")
-
-            .then()
-            .statusCode(HttpStatus.OK.value())
-            .body("id", equalTo(1))
-            .body("imageUrls", hasSize(2))
-            .body("imageUrls[0]", containsString("https://"))
-            .body("imageUrls[1]", containsString("https://"))
-            .body("writer", equalTo("정현석"))
-            .body("writerThumbnailImageUrl", equalTo(
-                "http://k.kakaocdn.net/dn/wtMIN/btrII2nrJAv/KWEi4dNNGqeBYjzr0KZGK1/img_110x110.jpg"))
-            .body("title", equalTo("강남역에서 떡볶이 먹을 사람 모집합니다."))
-            .body("location", equalTo("강남역"))
-            .body("latitude", equalTo(36.657677f))
-            .body("longitude", equalTo(128.300764f))
-            .body("description", equalTo("떡볶이 쉐어 설명"))
-            .body("price", equalTo(10000))
-            .body("originalPrice", equalTo(30000))
-            .body("currentRecruitment", equalTo(3))
-            .body("finalRecruitment", equalTo(3))
-            .body("recruitmentMemberThumbnailImageUrls", hasSize(3))
-            .body("recruitmentMemberThumbnailImageUrls[0]", containsString("http://"))
-            .body("recruitmentMemberThumbnailImageUrls[1]", containsString("http://"))
-            .body("createdDateTime", equalTo("2022-08-03 16:00"))
-            .body("closedDateTime", equalTo("2023-08-03 16:00"))
-            .body("wish", equalTo(false))
-            .body("entry", equalTo(false))
-            .body("wishCount", equalTo(1))
-            .body("locationNegotiation", equalTo(true))
-            .body("priceNegotiation", equalTo(false))
-            .body("locationGuide", equalTo("강남역 1번 출구"))
-            .body("hashtags", hasSize(2))
-            .body("hashtags[0]", equalTo("해시태그 내용 1"));
-    }
-
-    @Test
     void 특정_사용자가_작성한_쉐어를_조회한다() {
         given(documentationSpec)
             .param("writerId", 2370842997L)
